requirements.txt
fastapi
uvicorn
streamlit
pymupdf
google-genai
PyGithub
python-multipart
python-dotenv
requests
